added validatity checks

+++ src/controllers/taskController.js
@@ -3,6 +3,13 @@ const taskService = require('../services/taskService');
 async function create(req,res){
     try {
         const { title, description } = req.body;
+        
+        if (!title || title.trim() === "") {
+        return res.status(400).json({
+            success: false,
+            message: "Title is required"
+        });
+        }
 
         await taskService.createTask(title, description, req.user.id);
 
@@ -13,10 +20,17 @@ async function create(req,res){
 }
 
 async function remove(req,res){
-    try {
+    try { 
+        if (isNaN(req.params.id)) {
+            return res.status(400).json({
+                success: false,
+                message: "Invalid task ID"
+            });
+        }
         await taskService.deleteTask(req.params.id, req.user);
-
+        
         res.json({ success: true, message: "Task deleted" });
+        
     } catch (err) {
         res.status(500).json({ success: false, message: err.message });
     }
@@ -34,9 +48,25 @@ async function getAll(req,res){
 
 async function getSpecific(req,res){
     try {
+
+        if (isNaN(req.params.id)) {
+            return res.status(400).json({
+                success: false,
+                message: "Invalid task ID"
+            });
+        }
+        
         const tasks = await taskService.getTasks(req.params.id,req.user);
 
+        if (!task) {
+        return res.status(404).json({
+            success: false,
+            message: "Task not found"
+        });
+        }
+
         res.json({ success: true, data: tasks });
+
     } catch (err) {
         res.status(500).json({ success: false, message: err.message });
     }
@@ -45,6 +75,21 @@ async function getSpecific(req,res){
 async function update(req,res){
     try {
         const { title, description } = req.body;
+
+
+        if (!title || title.trim() === "") {
+        return res.status(400).json({
+            success: false,
+            message: "Title is required"
+        });
+        }
+
+        if (isNaN(req.params.id)) {
+        return res.status(400).json({
+            success: false,
+            message: "Invalid task ID"
+        });
+        }
 
         await taskService.updateTask(
         req.params.id,
